Rename useChartSynchroniser -> useSynchroniseCharts
